Uploaded 2/24/2022 scraped/feature-engineered data
If you are just interested in the resulting data, and don't want to run any programs. The final dataset "5a_AUF.csv" has been uploaded to GitHub. Keep in mind that this file is current up until 2/24/2022; fights which occur after this time will not be in "5a_AUF.csv", you'll need to run the programs to generate the most up to date dataset. Alternatively, if you do run the programs, then you can read in  5a_DataPlusHelperVars.R, and use the AUF dataframe.